DRY specs
- Add a JpiV1AdminController context for all the basic setup
- Reduce duplication between Dashboard and DashboardController
