+++ api/spec/controllers/mno_enterprise/jpi/v1/admin/impac/dashboard_templates_controller_spec.rb
@@ -1,75 +1,20 @@
 require 'rails_helper'
+require 'mno_enterprise/testing_support/shared_contexts/jpi_v1_admin_controller'
+require 'mno_enterprise/testing_support/shared_contexts/jpi_v1_admin_impac_controller'
 
 module MnoEnterprise
   describe Jpi::V1::Admin::Impac::DashboardTemplatesController, type: :controller do
-    # include MnoEnterprise::TestingSupport::JpiV1TestHelper
-    include MnoEnterprise::TestingSupport::SharedExamples::JpiV1Admin
-    render_views
-    routes { MnoEnterprise::Engine.routes }
-    before { request.env["HTTP_ACCEPT"] = 'application/json' }
+    include_context MnoEnterprise::Jpi::V1::Admin::BaseResourceController
 
-    RSpec.shared_context "#{described_class}: dashboard dependencies stubs" do
-      before do
-        api_stub_for(
-          get: "/users/#{user.id}/organizations",
-          response: from_api([org])
-        )
-        api_stub_for(
-          get: "/dashboards/#{template.id}/widgets",
-          response: from_api([widget])
-        )
-        api_stub_for(
-          get: "/dashboards/#{template.id}/kpis",
-          response: from_api([d_kpi])
-        )
-        api_stub_for(
-          get: "/widgets/#{widget.id}/kpis",
-          response: from_api([w_kpi])
-        )
-      end
+    include MnoEnterprise::TestingSupport::SharedContexts::JpiV1AdminImpacController
+    include_context 'MnoEnterprise::Jpi::V1::Admin::Impac' do
+      let(:dashboard) { template }
     end
 
-    let(:user) { build(:user, :admin, :with_organizations) }
-    let(:org) { build(:organization, users: [user]) }
-    let(:metadata) { { hist_parameters: { from: '2015-01-01', to: '2015-03-31', period: 'MONTHLY' } } }
     let(:template) { build(:impac_dashboard, dashboard_type: 'template', organization_ids: [org.uid], currency: 'EUR', settings: metadata, owner_type: nil, owner_id: nil, published: true) }
-    let(:widget) { build(:impac_widget, dashboard: template) }
-    let(:d_kpi) { build(:impac_kpi, dashboard: template) }
-    let(:w_kpi) { build(:impac_kpi, widget: widget) }
 
-    def hash_for_kpi(kpi)
-      {
-        "id" => kpi.id,
-        "element_watched" => kpi.element_watched,
-        "endpoint" => kpi.endpoint
-      }
-    end
-    let(:hash_for_widget) do
-      {
-        "id" => widget.id,
-        "name" => widget.name,
-        "endpoint" => widget.widget_category,
-        "width" => widget.width,
-        "kpis" => [hash_for_kpi(w_kpi)]
-      }
-    end
     let(:hash_for_template) do
-      {
-        "id" => template.id,
-        "name" => template.name,
-        "full_name" => template.full_name,
-        "currency" => 'EUR',
-        "metadata" => metadata.deep_stringify_keys,
-        "data_sources" => [{ "id" => org.id, "uid" => org.uid, "label" => org.name}],
-        "kpis" => [hash_for_kpi(d_kpi)],
-        "widgets" => [hash_for_widget],
-        "published" => true
-      }
-    end
-
-    before do
-      api_stub_for(get: "/users/#{user.id}", response: from_api(user))
-      sign_in user
+      hash_for_dashboard.merge("published" => true)
     end
       
     describe '#index' do
